Refactored functions, add functions with files

+++ include/calculator.h
@@ -1,11 +1,11 @@
 #ifndef CALCULATOR
 #define CALCULATOR
 
-void  calculator                       (char**,int);
-float calculate_expression             (float*, int*, char*, int*);
-float get_complete_integer             (char* src,int* j);
-float perform_operation                (float, float, char);
-void  process_high_priority_operations (float*, int*, char*, int*);
-void  process_low_priority_operations  (float*, int*, char*, int*);
+void   calculator                       (char**,int);
+double calculate_expression             (double*, int*, char*, int*);
+double get_complete_integer             (char* src,int* j);
+double perform_operation                (double, double, char);
+void   process_high_priority_operations (double*, int*, char*, int*);
+void   process_low_priority_operations  (double*, int*, char*, int*);
 
 #endif
